move set_pixel_as_point to hyp3_gamma.util
