+++ tests/test_util.py
@@ -1,6 +1,8 @@
 import os
 import shutil
 
+from osgeo import gdal
+
 from hyp3_gamma import util
 
 
@@ -38,3 +40,16 @@ def test_unzip_granule_and_remove(tmp_path, test_data_dir):
     assert safe_dir == 'S1A_IW_SLC__1SDV_20170525T025145_20170525T025157_016732_01BCA6_CEBE.SAFE'
     assert os.path.isdir(safe_dir)
     assert not os.path.exists(zip_file)
+
+
+def test_set_pixel_as_point(tmp_path, test_data_dir):
+    shutil.copy(test_data_dir / 'test_geotiff.tif', tmp_path)
+    geotiff = str(tmp_path / 'test_geotiff.tif')
+    info = gdal.Info(geotiff, format='json')
+    assert info['geoTransform'] == [440720.0, 60.0, 0.0, 3751320.0, 0.0, -60.0]
+    assert info['metadata']['']['AREA_OR_POINT'] == 'Area'
+
+    util.set_pixel_as_point(geotiff)
+    info = gdal.Info(geotiff, format='json')
+    assert info['geoTransform'] == [440750.0, 60.0, 0.0, 3751290.0, 0.0, -60.0]
+    assert info['metadata']['']['AREA_OR_POINT'] == 'Point'
